+++ packages/playground/src/environments/build.ts
@@ -1,6 +1,6 @@
 
 export const build = {
   version: '1.0.0',
-  revision: '87ee7f3',
-  date: '1609766899383'
+  revision: 'c0bb205',
+  date: '1609784933164'
 };