+++ packages/proxy/src/endpoints/proxyEndpoint.ts
@@ -5,15 +5,15 @@ export const proxyEndpoint = new class ProxyEndpoint {
   register(app: Express) {
 
     // todo add CSRF token
-    app.get('/api/proxy', (request: Request, response: Response) => {
-      proxyService.getProxiedHtml(request.query.url as string).then(proxyResponse => {
-        response.status(200);
-        response.send(proxyResponse.body);
-      }).catch(err => {
-        response.status(500);
-        response.json({error: true, message: err});
-        response.end();
-      });
-    });
+    // app.get('/api/proxy', (request: Request, response: Response) => {
+    //   proxyService.getProxiedHtml(request.query.url as string).then(proxyResponse => {
+    //     response.status(200);
+    //     response.send(proxyResponse.body);
+    //   }).catch(err => {
+    //     response.status(500);
+    //     response.json({error: true, message: err});
+    //     response.end();
+    //   });
+    // });
   }
 };